clean up, keeping just eventRegistry as before
